Add Timer and replacement uint64_t to bool

+++ src/Game/GameObject/Hero.h
@@ -25,7 +25,7 @@ public:
 		 std::shared_ptr<RenderEngine::Sprite> pSprite_bottom, 
 		 std::shared_ptr<RenderEngine::Sprite> pSprite_left, 
 		 std::shared_ptr<RenderEngine::Sprite> pSprite_right, 
-		 const float velocity, 
+		 const double velocity, 
 		 const glm::vec2& position, 
 		 const glm::vec2& size, 
 		 const float layer);
@@ -33,7 +33,7 @@ public:
 	void render() const override;
 	void SetOrientation(const EOrientaition eOrientation);
 	void move(const bool move);
-	void update(const uint64_t delta) override;
+	void update(const double delta) override;
 
 private:
 	EOrientaition m_eOrientation;
@@ -46,6 +46,6 @@ private:
 	RenderEngine::SpriteAnimator m_spriteAnimator_left;
 	RenderEngine::SpriteAnimator m_spriteAnimator_right;
 	bool m_move;
-	float m_velocity;
+	double m_velocity;
 	glm::vec2 m_moveOffSet;
 };
